mobile-guide/backend/therapy-exercises/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
PyJWT==2.8.0
python-dotenv==1.0.0
xgboost==2.0.3
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
networkx==3.2.1
experta==1.9.4
